requirements.txt
AMQPStorm==2.10.4
aniso8601==9.0.1
attrs==21.4.0
backports.zoneinfo==0.2.1
bson==0.5.10
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
dateparser==1.1.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
flask-restplus==0.13.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
pamqp==2.3.0
pika==1.2.0
PyJWT==2.3.0
pymongo==4.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.1.18
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
Werkzeug==0.16.1
zipp==3.6.0
urllib3==1.26.8
uWSGI==2.0.20
Faker